Move OpenNebula SDK related classes to a SDK package
