+++ src/main/java/cloud/fogbow/ras/core/plugins/interoperability/opennebula/sdk/v5_4/attachment/model/CreateAttachmentRequest.java
@@ -1,4 +1,4 @@
-package cloud.fogbow.ras.core.plugins.interoperability.opennebula.attachment.v5_4;
+package cloud.fogbow.ras.core.plugins.interoperability.opennebula.sdk.v5_4.attachment.model;
 
 public class CreateAttachmentRequest {
