+++ src/main/java/cloud/fogbow/ras/core/plugins/interoperability/opennebula/sdk/v5_4/network/model/VirtualNetworkUpdateTemplate.java
@@ -1,4 +1,4 @@
-package cloud.fogbow.ras.core.plugins.interoperability.opennebula.network.v5_4;
+package cloud.fogbow.ras.core.plugins.interoperability.opennebula.sdk.v5_4.network.model;
 
 import static cloud.fogbow.common.constants.OpenNebulaConstants.SECURITY_GROUPS;
 import static cloud.fogbow.common.constants.OpenNebulaConstants.TEMPLATE;
