+++ src/main/java/cloud/fogbow/ras/core/plugins/interoperability/opennebula/sdk/v5_4/securityrule/model/CreateSecurityGroupRequest.java
@@ -1,4 +1,4 @@
-package cloud.fogbow.ras.core.plugins.interoperability.opennebula.securityrule.v5_4;
+package cloud.fogbow.ras.core.plugins.interoperability.opennebula.sdk.v5_4.securityrule.model;
 
 import static cloud.fogbow.common.constants.OpenNebulaConstants.ID;
 import static cloud.fogbow.common.constants.OpenNebulaConstants.NAME;
